HireMe-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-docx==0.8.11
scikit-learn==1.3.0
nltk==3.8.1
pandas==2.0.3
numpy==1.24.3
python-magic==0.4.27
spacy==3.6.1 
